# Add grouped binary convolution support (3/3): indirect BGEMM kernel.