… larq_compute_engine/core/indirect_bgemm/kernel_8x4x2_aarch64.h
… larq_compute_engine/core/indirect_bgemm/kernel_8x4x4_aarch64.h
… larq_compute_engine/core/indirect_bgemm/prepare.h
… larq_compute_engine/core/indirect_bgemm/select_kernel.h
… larq_compute_engine/core/types.h
… larq_compute_engine/mlir/transforms/prepare_tf.cc
… larq_compute_engine/tflite/kernels/BUILD
… larq_compute_engine/tflite/kernels/bconv2d.cc
… larq_compute_engine/tflite/tests/bconv2d_test.cc