@AdamHillier 2020-11-06
> Awesome work! I am currious to see the benchmarks 🚀

So the gist of the benchmarks is that there are lovely speedups when using grouped convolutions, not much worse than the difference in MACs.

However, for non-grouped convolutions I'm recording a small but consistent slowdown, so I need to work out how to fix that before proceeding with this PR. I think it might require a bunch of templating that will ensure that certain instructions are only included for actual grouped convolutions.

I'll keep this a draft until I can resolve those issues.