@AdamHillier 2021-06-04
I've rebased this PR on top of main, incorporating the TF 2.5 changes.

### Assessing non-grouped performance

I'm not sure how, but now the performance regression I previously noticed on the Raspberry Pi 4B has disappeared!

These are measurements taken on my Raspberry Pi 4B (using non-grouped QuickNet models), showing the average latency from 100 runs and the std deviation. It looks like there *might* still be a slightly slow-down for the larger models, but it's very slight (previously I observed a 1-2 ms slowdown).

| Model          | Latency, indirect BGEMM baseline (ms) | Latency, indirect BGEMM this PR (ms) |
|----------------|---------------------------------------|--------------------------------------|
| QuickNet Small | 30.02 +- 0.19                         | 29.75 +- 0.13                        |
| QuickNet       | 48.06 +- 0.06                         | 48.37 +- 0.06                        |
| QuickNet Large | 78.92 +- 0.07                         | 79.36 +- 0.09                        |

In addition, per-op profiling now shows no significant difference between the latencies of the `ADD` ops, which was the concerning thing I previously saw.

### Assessing grouped performance

I generated a version of the `QuickNet` model used in the table above, but replaced every binary convolution with a grouped binary convolution with two groups. This means that the number of binary MACs is exactly half, giving a theoretical 2x speedup for the binary convolutions.

Here are benchmark results for both the whole model and just the binary convolutions (per-op benchmarks don't report std deviations), running on the same Raspberry Pi 4B device:

| Model                 | Latency, whole model (ms) | Latency, binary convolutions (ms) |
|-----------------------|---------------------------|-----------------------------------|
| QuickNet (no groups)  | 48.37 +- 0.06             | 27.86                             |
| QuickNet (two groups) | 36.51 +- 0.05             | 16.56                             |

There is a ~24% whole-model latency decrease, and ~41% binary convolution latency decrease.

Note that it wouldn't be possible to actually speed up the binary convolutions by 2x, because we still have to do the same number of output-transformations whether we use groups or not.